Replace `stream.ReadStream` with `async` iterator.
Replaced the `stream.ReadStream` with an `async` iterator. I was making
an effort to support Node.js 12 and Node.js 14. but my applications are
`async`/`await` and `async`/`await` behavior for `ReadStream` changed
between Node.js 12 and Node.js 14. If you are `async` iterating over
`ReadStream` in `destroy()` early in Node.js 12, the error is emitted as
an event, in Node.js 14 the error is raised as an exception. I did not
want to have a version dependent unit test on a library that had not yet
been released so either I drop support for Node.js 12 or use a different
interface.

If the user really wants to stream from a write-ahead log they can use
one of the shims on NPM. I don't know if they work, but that is not my
problem. They can antagonize the maintainer of an async iterator to
stream shim module instead of myself.

Also, I believe I'm going to drop suppport for Node.js 12 since Node.js
14 appears to resolve stack trace issues when programming with
`async`/`await` and I don't want to debug issues submitted with stubby
stack traces.