Add synchronization.
There will only ever be one writer strand for the write-ahead log. We
already manage read and write concurrency using the position index.
Readers make a copy of the latest index, a synchronous copy, so new
writes are not visible.

We needed to add a read/write lock to protect a log file when we shift
it. Was starting to add a readers count like in Amalgamate, but decided
to use Sequester. Already tested, less code. Did not write unit tests
for the concurrency, just looked at it really hard, the code, looked at
the code really, really hard.